+++ src/ui_bevy/ui_builder_react_ext.rs
@@ -45,13 +45,17 @@ pub trait UiBuilderReactExt
     fn despawn_on_broadcast<T: Send + Sync + 'static>(&mut self) -> &mut Self;
 
     /// Mirrors [`UiReactEntityCommandsExt::update_on`].
-    fn update_on<M, C, T>(&mut self, triggers: T, reactor: C) -> &mut Self
+    fn update_on<M, C, T, R>(&mut self, triggers: T, reactor: C) -> &mut Self
     where
         T: ReactionTriggerBundle,
-        C: IntoSystem<UpdateId, (), M> + Send + Sync + 'static;
+        R: ReactorResult,
+        C: IntoSystem<UpdateId, R, M> + Send + Sync + 'static;
 
     /// Mirrors [`UiReactEntityCommandsExt::update`].
-    fn update<M, C: IntoSystem<UpdateId, (), M> + Send + Sync + 'static>(&mut self, reactor: C) -> &mut Self;
+    fn update<M, R: ReactorResult, C: IntoSystem<UpdateId, R, M> + Send + Sync + 'static>(
+        &mut self,
+        reactor: C,
+    ) -> &mut Self;
 
     /// Mirrors [`UiReactEntityCommandsExt::modify`].
     fn modify(&mut self, callback: impl FnMut(EntityCommands) + Send + Sync + 'static) -> &mut Self;
@@ -92,16 +96,20 @@ impl UiBuilderReactExt for UiBuilder<'_, Entity>
         self
     }
 
-    fn update_on<M, C, T>(&mut self, triggers: T, reactor: C) -> &mut Self
+    fn update_on<M, C, T, R>(&mut self, triggers: T, reactor: C) -> &mut Self
     where
         T: ReactionTriggerBundle,
-        C: IntoSystem<UpdateId, (), M> + Send + Sync + 'static,
+        R: ReactorResult,
+        C: IntoSystem<UpdateId, R, M> + Send + Sync + 'static,
     {
         self.entity_commands().update_on(triggers, reactor);
         self
     }
 
-    fn update<M, C: IntoSystem<UpdateId, (), M> + Send + Sync + 'static>(&mut self, reactor: C) -> &mut Self
+    fn update<M, R: ReactorResult, C: IntoSystem<UpdateId, R, M> + Send + Sync + 'static>(
+        &mut self,
+        reactor: C,
+    ) -> &mut Self
     {
         self.entity_commands().update_on((), reactor);
         self