+++ CHANGELOG.md
@@ -21,9 +21,12 @@
     - Removed `WithAbsoluteNode` and `WithFlexNode`.
 - Add `SetClipMargin` loadable.
 - Re-enable `remove_with_requires` in instruction reversion now that the bevy bug is fixed.
-- Add tooltip widget.
+- Integrate `ReactorResult` from `bevy_cobweb`. All built-in callbacks like `.on_pressed` now let you early-out with `?`, so long as you return `Ok` (warns on error) or `DONE` (drops errors).
 - Bugfixes
     - Cursor compile error on WASM.
+
+- Add tooltip widget.
+
 
 ## 0.5.1